add dry run

+++ genie/build-root.sh
@@ -3,25 +3,32 @@
 # NAB:  Modified from /apps/root/6.26.10 on June 20, 2023
 #
 
-export ROOT_VERS=6.26.10
-export PREFIX=.
+VERSION=6.26.10
+PREFIX=.
+DRYRUN=0 
 
-while getopts "v:p:" opt
+while getopts "v:p:d" opt
 do
     case $opt in
-        v) ROOT_VERS=$OPTARG ;;
+        v) VERSION=$OPTARG ;;
         p) PREFIX=$(realpath $OPTARG) ;;
+        d) DRYRUN=1 ;;
         ?) echo "Usage: build-root.sh [-v version] [-p prefix]" && exit 1 ;;
     esac
 done
 
-export GCC_VERS=$(gcc --version | head -1 | awk '{print$3}')
 export CXX=`which g++`
 export CC=`which gcc`
+STUB=$(gcc --version | head -1 | awk '{print$3}')
 
-STUB=${ROOT_VERS}-gcc${GCC_VERS}
+if [ $DRYRUN -ne 0 ]
+then
+    echo ROOT VERSION:       $VERSION
+    echo ROOT INSTALL DIR:   $PREFIX/$STUB
+    exit
+fi
 
-[ -e ${ROOT_VERS}.src ] && echo ERROR:  source directory already exists: root-${ROOT_VERS}.src && exit 1
+[ -e ${VERSION}.src ] && echo ERROR:  source directory already exists: root-${VERSION}.src && exit 1
 [ -e ${STUB}.build ] && echo ERROR:  build directory already exists: ${STUB}.build && exit 2
 [ -e ${PREFIX}/${STUB} ] && echo ERROR:  install directory already exists: ${PREFIX}/${STUB} && exit 3
 
@@ -31,13 +38,13 @@ touch ${STUB}.log
 set -x -e
 
 # Download:
-wget https://root.cern/download/root_v${ROOT_VERS}.source.tar.gz |& tee -a ${STUB}.log
-tar xzf root_v${ROOT_VERS}.source.tar.gz |& tee -a ${STUB}.log
-mv root-${ROOT_VERS} root-${ROOT_VERS}.src |& tee -a ${STUB}.log
+wget https://root.cern/download/root_v${VERSION}.source.tar.gz |& tee -a ${STUB}.log
+tar xzf root_v${VERSION}.source.tar.gz |& tee -a ${STUB}.log
+mv root-${VERSION} root-${VERSION}.src |& tee -a ${STUB}.log
 
 # Configure:
 d=$(dirname $(realpath ${BASH_SOURCE[0]}))
-cmake -S root-${ROOT_VERS}.src -B ${STUB}.build \
+cmake -S root-${VERSION}.src -B ${STUB}.build \
  -DCMAKE_CXX_STANDARD=17 -DCMAKE_INSTALL_PREFIX=${PREFIX}/${STUB} -Dbuiltin_glew=ON \
  -Dpythia6=ON -DPYTHIA6_LIBRARY=$d/lib/libPythia6.so \
  |& tee -a ${STUB}.log
